Added titles when mouse hovers over a YouTube link
The youtube oembed API is compatible with the most recent YouTube API changes, and requires no authentication (unlike the v3 API).  This method uses the content script to retrieve title information from the YouTube API, bypassing the Access-Control-Allow-Origin issue, and requiring no middleman server (it gives permission to the extension for YouTube URLs).  However, unless someone else implements a way to do the same for the other browsers' extension, it would be Chrome only.  Also, since the main betterttv.js would still propagate to all browsers and base versions, it would need to check that it's running in Chrome and on the latest extension version.  Other than that, it's been working flawlessly for me in Chrome in my testing, and works for all types of YouTube video URLs I've encountered, as well as YouTube playlist URLs.