Add the ability to disable  player gradient overlays.
When the mouse hovers over the player a gradient overlays the video
at the bottom.  Depending on your player size this gradient can
can actually overlay a large portion of the video and make it
particularly difficult to read any text the streamer has at the
bottom of their screen.