Added /follow and /unfollow commands
After night closed issue #2606 and @Stoupedog made it seem like i'm having the same issue as in the issue that is in #2019 I figured out i can take couple nights and learn JS and add /follow and /unfollow commands to channel. My biggest argument is there is more convenient for most people I asked around my friends and they agreed that it would be really useful to have those commands. Second argument is that some people have automatically turned on Theater mode (Function of BTTV!) and they may like to follow by just typing in command. Well at the end it's quality of life thing so it's not number one priority but imo it could help to alot of people out of laziness to not follow the streamer and support more streamers that not get much support.