Added a few options for the sidebar
ported to new-twitch branch from
https://github.com/night/BetterTTV/pull/2647/files.

Also jquery is cool
and already imported so why not add an option to auto expand followed
channels as well.

The hiding of offline channels that are followed is
pretty annoying now though since you have to use javascript to get the
parent elements. I'm not a fan of the approach I used as it can bug out
when enabled before expanding but will auto correct on frame resizes
(another thing I'm not a fan of).